new test syscall

+++ linux/include/linux/syscalls.h
@@ -832,7 +832,7 @@ asmlinkage long sys_fanotify_mark(int fanotify_fd, unsigned int flags,
 asmlinkage long sys_syncfs(int fd);
 
 /* e6998 */
-asmlinkage long sys_vcpu_info(void);
+asmlinkage int sys_vcpu_info(void);
 
 
 asmlinkage long sys_fork(void);
